+++ src/actors/circle.rs
@@ -108,6 +108,7 @@ impl Drawable for Circle{
                 return true
             }
         }
+        
         return false
     }
 }